Project files cleanup
