+++ Screens/Focus.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { View, Text } from "react-native";
-import Timer from "../components/Timer";
+import Session from "../components/Session";
 
 function Focus(props) {
   return (
     <View>
-      <Timer />
+      <Session />
     </View>
   );
 }
